[BE] REFACTOR: 조건 수정에 따른 service, route 수정

+++ server/src/services/task.js
@@ -6,22 +6,28 @@ const errorCode = require('@utils/error-codes');
 const { models } = sequelize;
 const taskModel = models.task;
 
+// const isValidTask = task => {
+//   if (!task) {
+//     const error = new Error(errorMessage.NOT_FOUND_ERROR('task'));
+//     error.status = errorCode.NOT_FOUND_ERROR;
+//     throw error;
+//   }
+// };
+
 const retrieveById = async ({ id, userId }) => {
   const task = await taskModel.findByPk(id, {
     include: [
-      'labels',
-      'priority',
-      'alarm',
       'bookmarks',
       {
         model: taskModel,
-        include: ['labels', 'priority', 'alarm', 'bookmarks'],
+        include: ['bookmarks'],
       },
     ],
     order: [[taskModel, 'position', 'ASC']],
   });
 
   // task가 없는 경우, url params로 넘어오는 taskId가 유효하지 않음
+  // isValidTask(task);
   if (!task) {
     const error = new Error(errorMessage.NOT_FOUND_ERROR('task'));
     error.status = errorCode.NOT_FOUND_ERROR;
@@ -30,7 +36,7 @@ const retrieveById = async ({ id, userId }) => {
 
   // 요청받은 task가 해당 유저의 작업이 아닌 경우 리소스 접근 권한이 없음
   if (!(await isTaskOwner({ id, userId }))) {
-    const error = new Error(errorMessage.FORBIDDEN_ERROR);
+    const error = new Error(errorMessage.FORBIDDEN_ERROR('task'));
     error.status = errorCode.FORBIDDEN_ERROR;
     throw error;
   }
@@ -39,20 +45,24 @@ const retrieveById = async ({ id, userId }) => {
 
 const retrieveAll = async userId => {
   const task = await taskModel.findAll({
-    where: { isDone: false },
     include: [
-      'labels',
-      'priority',
-      'alarm',
       'bookmarks',
       {
         model: taskModel,
-        include: ['labels', 'priority', 'alarm', 'bookmarks'],
+        include: ['bookmarks'],
+        where: { isDone: false },
       },
       {
-        model: models.project,
-        attributes: [],
-        where: { creatorId: userId },
+        model: models.section,
+        attribute: [],
+        include: [
+          {
+            model: models.project,
+            attributes: [],
+            where: { creatorId: userId },
+          },
+        ],
+        required: true,
       },
     ],
     order: [[taskModel, 'position', 'ASC']],
@@ -61,7 +71,7 @@ const retrieveAll = async userId => {
 };
 
 const create = async ({ projectId, sectionId, userId, ...taskData }) => {
-  const { labelIdList, dueDate, ...rest } = taskData;
+  const { dueDate, ...rest } = taskData;
 
   const project = await models.project.findByPk(projectId);
   if (!project) {
@@ -70,7 +80,7 @@ const create = async ({ projectId, sectionId, userId, ...taskData }) => {
     throw error;
   }
   if (!(await isProjectOwner({ id: projectId, userId }))) {
-    const error = new Error(errorMessage.FORBIDDEN_ERROR);
+    const error = new Error(errorMessage.FORBIDDEN_ERROR('project'));
     error.status = errorCode.FORBIDDEN_ERROR;
     throw error;
   }
@@ -84,7 +94,7 @@ const create = async ({ projectId, sectionId, userId, ...taskData }) => {
     }
     if (!(await isSectionOwner({ id: sectionId, userId }))) {
       const error = new Error(errorMessage.FORBIDDEN_ERROR('section'));
-      error.status = errorCode.NOT_FOUND_ERROR;
+      error.status = errorCode.FORBIDDEN_ERROR;
       throw error;
     }
 
@@ -99,12 +109,9 @@ const create = async ({ projectId, sectionId, userId, ...taskData }) => {
     }, 0);
 
     const task = await models.task.create(
-      { projectId, sectionId, dueDate, position: maxPosition + 1, ...rest },
+      { sectionId, dueDate, position: maxPosition + 1, ...rest },
       { transaction: t },
     );
-    if (labelIdList) {
-      await task.setLabels(JSON.parse(labelIdList), { transaction: t });
-    }
 
     return task;
   });
@@ -113,7 +120,7 @@ const create = async ({ projectId, sectionId, userId, ...taskData }) => {
 };
 
 const update = async taskData => {
-  const { id, labelIdList, dueDate, userId, ...rest } = taskData;
+  const { id, dueDate, userId, ...rest } = taskData;
 
   const result = await sequelize.transaction(async t => {
     try {
@@ -126,15 +133,13 @@ const update = async taskData => {
 
       // 요청받은 task가 해당 유저의 작업이 아닌 경우 리소스 접근 권한이 없음
       if (!(await isTaskOwner({ id, userId }))) {
-        const error = new Error(errorMessage.FORBIDDEN_ERROR);
+        const error = new Error(errorMessage.FORBIDDEN_ERROR('task'));
         error.status = errorCode.FORBIDDEN_ERROR;
         throw error;
       }
 
       await task.update({ dueDate, ...rest });
-      if (labelIdList) {
-        await task.setLabels(JSON.parse(labelIdList), { transaction: t });
-      }
+
       task.save();
       return true;
     } catch (err) {
@@ -147,6 +152,12 @@ const update = async taskData => {
 };
 
 const remove = async id => {
+  const task = await taskModel.findByPk(id);
+  if (!task) {
+    const error = new Error(errorMessage.NOT_FOUND_ERROR('task'));
+    error.status = errorCode.NOT_FOUND_ERROR;
+    throw error;
+  }
   const result = await taskModel.destroy({
     where: {
       id,
